# apps/leaderboard-client/src/app/api/openapi.yaml
openapi: 3.0.3
info:
  title: Leaderboard API
  description: |
    API REST du projet Leaderboard — gestion de projets, challenges, tâches, contributions, évaluations et sync meetings.

    ## Authentification
    L'API utilise des cookies HTTP-only (`access_token` / `refresh_token`) via OAuth2 Google.
    Certaines routes nécessitent le rôle **admin**.
  version: 1.0.0
servers:
  - url: /api
    description: Next.js API routes
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
    cronBearer:
      type: http
      scheme: bearer
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
    ValidationError:
      type: object
      properties:
        error:
          type: string
          example: Validation error
        details:
          type: array
          items:
            type: object
      required: [error, details]
    Success:
      type: object
      properties:
        success:
          type: boolean
          example: true

    # ── Users ──
...
    Challenge:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        title:
          type: string
        slug:
          type: string
          description: Segment of the public page URL, /challenges/{slug}. Changing it keeps the old one as a redirect.
        status:
          type: string
        start_date:
          type: string
          format: date-time
        end_date:
          type: string
          format: date-time
        description:
          type: string
          nullable: true
        roadmap:
          type: string
          nullable: true
        completion:
          type: number
        contribution_points_reward:
          type: integer
        project_id:
          type: string
          format: uuid
    CreateChallenge:
      type: object
      required: [title, status, start_date, end_date, contribution_points_reward, project_id]
      properties:
        title:
          type: string
          minLength: 1
        slug:
          type: string
          pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$'
          minLength: 3
          maxLength: 80
          description: Optional. Derived from the title when omitted; 409 when taken.
        status:
          type: string
        start_date:
          type: string
          format: date-time
        end_date:
          type: string
          format: date-time
        description:
          type: string
        roadmap:
          type: string
        contribution_points_reward:
          type: integer
          minimum: 0
        project_id:
          type: string
          format: uuid
    UpdateChallenge:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        slug:
          type: string
          pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$'
          minLength: 3
          maxLength: 80
          description: Changing it keeps the previous slug as a redirect; 409 when taken.
        status:
          type: string
        start_date:
          type: string
          format: date-time
        end_date:
          type: string
          format: date-time
        description:
          type: string
        roadmap:
          type: string
        contribution_points_reward:
          type: integer
          minimum: 0
        project_id:
          type: string
          format: uuid
    SlugAvailability:
      type: object
      properties:
        slug:
          type: string
        available:
          type: boolean
        problem:
          type: string
          nullable: true
          description: Why the slug cannot be used, as a sentence to display.
        suggestion:
          type: string
          nullable: true
          description: A free slug to offer instead, when available is false.
    SlugTaken:
      type: object
      properties:
        error:
          type: string
        field:
          type: string
          enum: [slug]
        suggestion:
          type: string

    # ── Tasks ──
...
paths:
...
  /challenges:
    get:
      tags: [Challenges]
      summary: Lister tous les challenges
      responses:
        '200':
          description: Liste des challenges
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Challenge'
    post:
      tags: [Challenges]
      summary: Créer un challenge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChallenge'
      responses:
        '201':
          description: Challenge créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Challenge'
        '400':
          description: Erreur de validation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '409':
          description: Slug déjà pris
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SlugTaken'
  /challenges/slug-availability:
    get:
      tags: [Challenges]
      summary: Vérifier qu'un slug de challenge est libre
      description: Admins et managers de projet. Ne réserve rien — l'index unique tranche à l'écriture.
      parameters:
        - name: slug
          in: query
          required: true
          schema:
            type: string
        - name: exclude
          in: query
          required: false
          description: Le challenge en cours d'édition, dont le slug et les anciens slugs restent disponibles.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Disponibilité du slug
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SlugAvailability'
        '401':
          description: Non authentifié
        '403':
          description: Ni admin ni manager
  /challenges/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Challenges]
      summary: Récupérer un challenge
      responses:
        '200':
          description: Challenge trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Challenge'
        '404':
          description: Challenge non trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags: [Challenges]
      summary: Mettre à jour un challenge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateChallenge'
      responses:
        '200':
          description: Challenge mis à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Challenge'
        '400':
          description: Erreur de validation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
    delete:
      tags: [Challenges]
      summary: Supprimer un challenge
      responses:
        '200':
          description: Challenge supprimé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
...